Move unified about logic to view model
