+++ WordPress/src/main/java/org/wordpress/android/ui/about/UnifiedAboutNavigationAction.kt
@@ -1,0 +1,5 @@
+package org.wordpress.android.ui.about
+
+sealed class UnifiedAboutNavigationAction {
+    object Dismiss : UnifiedAboutNavigationAction()
+}
